# src/main/java/com/dodibo/learncore/auth/dto/ResendActivationRequest.java
package com.dodibo.learncore.auth.dto;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class ResendActivationRequest {

    @NotBlank
    @Email
    private String email;
}