# src/main/java/com/dodibo/learncore/auth/dto/ResetPasswordRequest.java
package com.dodibo.learncore.auth.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class ResetPasswordRequest {

    @NotBlank
    private String token;

    @NotBlank
    @Size(min = 8, message = "Password should be 8 character minimum")
    private String newPassword;
}